+++ CHANGELOG.md
@@ -16,6 +16,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - Overlay can be moved more precisely
   - Dungeons can be "joined" and "left" manually (if the automatic detection fails): `/moo dungeon <enter/leave>`
 - Improved handling of invalid/missing Hypixel API key
+- `/moo stalkskyblock` Switched from sky.lea.moe (discontinued) to sky.shiiyu.moe
+
+### Fixed
+- Fixed crash caused by another, outdated and buggy mod which sadly too many people still use
+- various smaller fixes here and there, e.g.:
+  - 'Create Auction' and 'Create BIN Auction' now show the price per item if multiple items are sold
+  - Dungeon party finder: entered vs queued floor wasn't detected correctly
+  - A dead player was counted as another death when they left the SkyBlock dungeon
 
 ## [1.8.9-0.10.2] - 15.09.2020
 ### Added